only expose display_data in rg15.h
