+++ src/sensors/rg15.c
@@ -10,6 +10,16 @@
 #define DEFAULT_DEVICE "/dev/ttyUSB0"
 #define DARWIN_DEVICE "/dev/tty.usbserial-0001"
 
+typedef struct {
+    char key[32];
+    char value[32];
+} KeyValuePair;
+
+typedef struct {
+    int fd;         // File descriptor
+    char device[64];
+    struct termios tty;
+} RG15Device;
 // Initialize the serial device
 RG15Device* rg15_init(const char* device) {
     RG15Device* dev = malloc(sizeof(RG15Device));
